Memory
This commit introduces C++14 feature of std::make_unique to C++11 by providing a custom implementation based on the proposed standard.
All files that used to include <memory> now include "memory.hpp", which in turn includes <memory>, so any change affects only implementations that leverage the new std::make_unique, provided that the C++ version in use is not C++14 or above.