Generic interpolation
This commit introduces a new interpolation mechanism that can be used to interpolate a function starting from an arbitrary number of points.
For now, only two points are used in the implementation of the approximate power model (hence the implementation is equivalent to the previous one).
However, if needed this number can be easily increased leveraging the functions in the "interpolate.hpp" file.